update: template  constants moved to root constants
